CNDB-16344: SAI Rerankless ANN needs FP sim. score for coordinator's sort (#2184)
### What is the issue
Fixes: https://github.com/riptano/cndb/issues/16344

### What does this PR fix and why was it fixed
When `ANN_USE_SYNTHETIC_SCORE` is true, we send the score from the
replica to the coordinator. We need that score to be the full precision
score to ensure correct ordering. When a query supplies `rerank_k: 0`,
the score is approximate though and can produce out of order results, as
the test demonstrates. The fix is to configure the
`NodeQueueRowIdIterator` appropriately to indicate when the score is in
fact approximate. The FP score is only computed when
`ANN_USE_SYNTHETIC_SCORE` is true.

+++ src/java/org/apache/cassandra/cql3/UntypedResultSet.java
@@ -391,7 +391,7 @@ public abstract class UntypedResultSet implements Iterable<UntypedResultSet.Row>
             return DoubleType.instance.compose(data.get(column));
         }
 
-        public double getFloat(String column)
+        public float getFloat(String column)
         {
             return FloatType.instance.compose(data.get(column));
         }
